feat(init) : 초기 개발 세팅
