+++ .gitignore
@@ -1,24 +1,33 @@
+########################################
+# Project / Build
+########################################
 HELP.md
-.gradle
+.gradle/
 build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
 
-### STS ###
+
+########################################
+# STS (Spring Tool Suite)
+########################################
 .apt_generated
 .classpath
 .factorypath
 .project
-.settings
+.settings/
 .springBeans
-.sts4-cache
+.sts4-cache/
 bin/
 !**/src/main/**/bin/
 !**/src/test/**/bin/
 
-### IntelliJ IDEA ###
-.idea
+
+########################################
+# IntelliJ IDEA
+########################################
+.idea/
 *.iws
 *.iml
 *.ipr
@@ -26,16 +35,26 @@ out/
 !**/src/main/**/out/
 !**/src/test/**/out/
 
-### NetBeans ###
+
+########################################
+# NetBeans
+########################################
 /nbproject/private/
 /nbbuild/
 /dist/
 /nbdist/
 /.nb-gradle/
 
-### VS Code ###
+
+########################################
+# VS Code
+########################################
 .vscode/
 
-### Environment & Config files ###
+
+########################################
+# Environment & Config (IMPORTANT)
+########################################
 .env
-src/main/resources/*.yml
+src/main/resources/application-local.yml
+src/main/resources/application-prod.yml
