+++ build.gradle
@@ -25,35 +25,63 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    compileOnly 'org.projectlombok:lombok'
-    developmentOnly 'org.springframework.boot:spring-boot-devtools'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
+    /* =====================================================
+     * Spring Boot Core
+     * ===================================================== */
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+    /* =====================================================
+     * Persistence (JPA / Database)
+     * ===================================================== */
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.mysql:mysql-connector-j'
+
+    /* =====================================================
+     * View Template (Thymeleaf)
+     * ===================================================== */
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
     implementation 'nz.net.ultraq.thymeleaf:thymeleaf-layout-dialect'
+    implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
 
+    /* =====================================================
+     * Security
+     * ===================================================== */
+    implementation 'org.springframework.boot:spring-boot-starter-security'
 
+    /* =====================================================
+     * JWT Authentication
+     * ===================================================== */
+    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+
+    /* =====================================================
+     * API Documentation (Swagger / OpenAPI)
+     * ===================================================== */
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.13'
 
+    /* =====================================================
+     * Logging
+     * ===================================================== */
     implementation 'org.bgee.log4jdbc-log4j2:log4jdbc-log4j2-jdbc4.1:1.16'
 
-//시큐리티 추가
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
+    /* =====================================================
+     * Lombok
+     * ===================================================== */
+    compileOnly 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
+
+    /* =====================================================
+     * Development & Testing
+     * ===================================================== */
+    developmentOnly 'org.springframework.boot:spring-boot-devtools'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
-
-
-// JWT 관련 라이브러리
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'  // JWT API
-    implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'  // JWT 구현체
-    implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5'  // JWT에 Jackson을 사용하여 JSON 처리
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
+
 
 tasks.named('test') {
     useJUnitPlatform()
